+++ src/middlewares/timeout.js
@@ -48,8 +48,9 @@ module.exports = function(broker) {
 	}
 
 	return {
-		created(broker) {
+		name: "Timeout",
 
+		created(broker) {
 			if (broker.isMetricsEnabled()) {
 				broker.metrics.register({ name: METRIC.MOLECULER_REQUEST_TIMEOUT_TOTAL, type: METRIC.TYPE_COUNTER, labelNames: ["action"] });
 			}