+++ test/integration/middlewares.spec.js
@@ -38,7 +38,7 @@ describe("Test middleware system", () => {
 			}
 		};
 
-		let broker = new ServiceBroker({ logger: false, validation: false, internalMiddlewares: false, middlewares: [mw2Async, mw1Sync] });
+		let broker = new ServiceBroker({ logger: false, validator: false, internalMiddlewares: false, middlewares: [mw2Async, mw1Sync] });
 
 		let master = jest.fn(() => {
 			flow.push("MASTER");
@@ -112,7 +112,7 @@ describe("Test middleware system", () => {
 			return { user: "icebob" };
 		});
 
-		let broker = new ServiceBroker({ logger: false, validation: false, internalMiddlewares: false, middlewares: [mw3, mw2, mw1] });
+		let broker = new ServiceBroker({ logger: false, validator: false, internalMiddlewares: false, middlewares: [mw3, mw2, mw1] });
 
 		broker.createService({
 			name: "test",
@@ -184,7 +184,7 @@ describe("Test middleware system", () => {
 			return { user: "icebob" };
 		});
 
-		let broker = new ServiceBroker({ logger: false, validation: false, internalMiddlewares: false, middlewares: [mw3, mw2, mw1] });
+		let broker = new ServiceBroker({ logger: false, validator: false, internalMiddlewares: false, middlewares: [mw3, mw2, mw1] });
 
 		broker.createService({
 			name: "test",
@@ -241,7 +241,7 @@ describe("Test middleware system", () => {
 			});
 		});
 
-		let broker = new ServiceBroker({ logger: false, validation: false, middlewares: [mw2, mw1] });
+		let broker = new ServiceBroker({ logger: false, validator: false, middlewares: [mw2, mw1] });
 
 		broker.createService({
 			name: "test",